Add fields to indicate whether a robot should move/rotate
This is useful when you do want to send commands to the robot, but it should not move/rotate. For example, when it is already at the right position, or it should only kick. Without this, the robot would always try to move closer to its target angle/velocity, even if it is essentially already there. This results in the robot stuttering, which looks funky and increases wear and battery use.